2024: Keep in sync

+++ lib/custom.d/generic/rpi/45-mesa-vc4.sh
@@ -5,43 +5,23 @@
 MESA_REPO_URL="https://gitlab.freedesktop.org/mesa/mesa.git"
 MESA_BRANCH="main"
 # Remove tag if you want to build code from the branch above
-MESA_TAG="mesa-22.3.4"
+MESA_TAG="mesa-24.1.0"
 MESA_SRC_DIR=${EXTRADIR}/mesa
 MESA_OUT_DIR=${MESA_SRC_DIR}/build/${LINUX_PLATFORM}-vc4
 MESA_PREFIX=/usr
-MESA_FORCE_UPDATE=${MESA_FORCE_UPDATE:="yes"}
+MESA_FORCE_UPDATE=${MESA_FORCE_UPDATE:="no"}
 MESA_FORCE_REBUILD=${MESA_FORCE_REBUILD:="yes"}
 
-MESA_CROSS_PKGCONFIG="${MESA_OUT_DIR}/cross-pkg-config.sh"
-
-MESA_GCC="${MESA_CROSS_COMPILE}gcc"
-MESA_CXX="${MESA_CROSS_COMPILE}g++"
-MESA_AR="${MESA_CROSS_COMPILE}ar"
-MESA_STRIP="${MESA_CROSS_COMPILE}strip"
-MESA_NM="${MESA_CROSS_COMPILE}nm"
 
 # -----------------------------------------------------------------------------
 
 mesa_cross_init()
 {
-	cat <<-EOF > ${MESA_CROSS_PKGCONFIG}
-#!/bin/sh
-
-SYSROOT=${SYSROOT_DIR}
-
-export PKG_CONFIG_DIR=
-export PKG_CONFIG_SYSROOT_DIR=\${SYSROOT}
-export PKG_CONFIG_LIBDIR=\${SYSROOT}/usr/lib/${LINUX_PLATFORM}/pkgconfig:\${SYSROOT}/usr/lib/pkgconfig:\${SYSROOT}/usr/share/pkgconfig
-
-exec pkg-config "\$@"
-EOF
-	chmod +x ${MESA_CROSS_PKGCONFIG}
-
 	cat <<-EOF > ${MESA_OUT_DIR}/${MESON_CROSSFILE}
 # Meson cross-file for Mesa 3D Graphics Library, ver ${MESA_DEB_VER}
 [constants]
-compile_flags = []
-c_flags = [ '-I${SYSROOT_DIR}/usr/include', '-I${SYSROOT_DIR}/usr/include/${LINUX_PLATFORM}' ]
+common_flags = [ '--sysroot=${SYSROOT_DIR}', '-I${SYSROOT_DIR}/usr/include', '-I${SYSROOT_DIR}/usr/include/${LINUX_PLATFORM}' ]
+c_flags = []
 cpp_flags = []
 
 [binaries]
@@ -50,12 +30,16 @@ cpp = '${MESA_CXX}'
 ar = '${MESA_AR}'
 strip = '${MESA_STRIP}'
 nm = '${MESA_NM}'
-pkgconfig = '${MESA_CROSS_PKGCONFIG}'
-#exe_wrapper = 'QEMU_LD_PREFIX=${SYSROOT_DIR} ${QEMU_BINARY}'
+pkg-config = '/usr/bin/pkg-config'
+# exe_wrapper = 'QEMU_LD_PREFIX=${SYSROOT_DIR} ${QEMU_BINARY}'
 
 [built-in options]
-c_args = compile_flags + c_flags
-cpp_args = compile_flags + cpp_flags
+c_args = common_flags + c_flags
+cpp_args = common_flags + cpp_flags
+
+[properties]
+sys_root = '${SYSROOT_DIR}'
+pkg_config_libdir = '${SYSROOT_DIR}/usr/lib/${LINUX_PLATFORM}/pkgconfig:/usr/lib/pkgconfig:/usr/share/pkgconfig'
 
 [host_machine]
 system = 'linux'
@@ -74,32 +58,32 @@ mesa_prepare()
 		rm -rf $MESA_SRC_DIR
 	fi
 
-	if [ -d $MESA_SRC_DIR ] && [ -d ${MESA_SRC_DIR}/.git ] ; then
-		local old_url=$(git -C $MESA_SRC_DIR config --get remote.origin.url)
+	if [ -d ${MESA_SRC_DIR} ] && [ -d ${MESA_SRC_DIR}/.git ] ; then
+		local old_url=$(git -C ${MESA_SRC_DIR} config --get remote.origin.url)
 		if [ "${old_url}" != "${MESA_REPO_URL}" ] ; then
-			rm -rf $MESA_SRC_DIR
+			rm -rf ${MESA_SRC_DIR}
 		fi
 	fi
-	if [ -d $MESA_SRC_DIR ] && [ -d ${MESA_SRC_DIR}/.git ] ; then
+	if [ -d ${MESA_SRC_DIR} ] && [ -d ${MESA_SRC_DIR}/.git ] ; then
 		# update sources
-		git -C $MESA_SRC_DIR fetch origin --tags
+		git -C ${MESA_SRC_DIR} fetch origin --tags
 
-		git -C $MESA_SRC_DIR reset --hard
-		git -C $MESA_SRC_DIR clean -fdx
+		git -C ${MESA_SRC_DIR} reset --hard
+		git -C ${MESA_SRC_DIR} clean -fdx
 
 		echo "Checking out branch: ${MESA_BRANCH}"
-		git -C $MESA_SRC_DIR checkout -B $MESA_BRANCH origin/${MESA_BRANCH}
-		git -C $MESA_SRC_DIR pull
+		git -C ${MESA_SRC_DIR} checkout -B ${MESA_BRANCH} origin/${MESA_BRANCH}
+		git -C ${MESA_SRC_DIR} pull
 	else
-		[[ -d $MESA_SRC_DIR ]] && rm -rf $MESA_SRC_DIR
+		[[ -d ${MESA_SRC_DIR} ]] && rm -rf ${MESA_SRC_DIR}
 
 		# clone sources
-		git clone $MESA_REPO_URL -b $MESA_BRANCH $MESA_SRC_DIR
+		git clone ${MESA_REPO_URL} -b ${MESA_BRANCH} --tags ${MESA_SRC_DIR}
 	fi
 
 	if [ -n "${MESA_TAG}" ] ; then
 		echo "Checking out tag: tags/${MESA_TAG}"
-		git -C $MESA_SRC_DIR checkout tags/${MESA_TAG}
+		git -C ${MESA_SRC_DIR} checkout tags/${MESA_TAG}
 
 		MESA_RELEASE=$(head -n 1 ${MESA_SRC_DIR}/VERSION)
 		MESA_DEB_VER="${MESA_RELEASE}-tag"
@@ -109,8 +93,7 @@ mesa_prepare()
 		MESA_DEB_VER="${MESA_RELEASE}-${LAST_COMMIT_ID}"
 	fi
 
-	MESA_GCC_VER=$(${MESA_GCC} -dumpversion)
-	MESA_DEB_PKG_VER="${MESA_DEB_VER}-${DEBIAN_RELEASE}-${DEBIAN_RELEASE_ARCH}-${SOC_FAMILY}-${MESA_TOOLCHAIN}${MESA_GCC_VER}"
+	MESA_DEB_PKG_VER="${MESA_DEB_VER}-${DEBIAN_RELEASE}-${DEBIAN_RELEASE_ARCH}-${SOC_FAMILY}"
 	MESA_DEB_PKG="mesa-${MESA_DEB_PKG_VER}"
 	MESA_DEB_DIR="${DEBS_DIR}/${MESA_DEB_PKG}-deb"
 
@@ -119,25 +102,27 @@ mesa_prepare()
 
 mesa_make()
 {
+	mkdir -p ${MESA_OUT_DIR}
+
 	if [ "${MESA_FORCE_REBUILD}" = yes ] ; then
 		echo "Force rebuild"
-		rm -rf ${MESA_OUT_DIR}
+		rm -rf ${MESA_OUT_DIR}/*
 	fi
-	mkdir -p ${MESA_OUT_DIR}
 
 	mesa_cross_init
 
+	cd ${MESA_SRC_DIR}
+
 	echo "Configure Mesa..."
 
-	cd ${MESA_OUT_DIR}
-
-	${MESON_DIR}/meson.py ${MESA_SRC_DIR}/ --cross-file="${MESON_CROSSFILE}" \
+	${MESON_DIR}/meson.py setup ${MESA_OUT_DIR}/ \
+			--cross-file="${MESA_OUT_DIR}/${MESON_CROSSFILE}" \
 			--prefix="${MESA_PREFIX}" \
 			--errorlogs \
 			--backend=ninja \
 			-Dplatforms= \
-			-Dgallium-drivers=vc4,kmsro,swrast \
-			-Ddri-drivers= \
+			-Dgallium-drivers=vc4,v3d,kmsro,swrast \
+			-Degl-native-platform=drm \
 			-Dvulkan-drivers= \
 			-Dgles2=enabled \
 			-Degl=enabled \
@@ -147,16 +132,27 @@ mesa_make()
 			-Dlibunwind=disabled \
 			-Dgallium-vdpau=disabled
 
+	[ $? -eq 0 ] || exit 1
+
 	echo "Making Mesa..."
 
-	ninja -v
+        ${MESON_DIR}/meson.py compile -C ${MESA_OUT_DIR}/
 
-	DESTDIR="./dist" ninja install
+	[ $? -eq 0 ] || exit 1
 
-	# NOTE: only need vc4-drm, so remove other libraries
+        DESTDIR="${MESA_OUT_DIR}/dist" \
+                ${MESON_DIR}/meson.py install -C ${MESA_OUT_DIR}/
+
+	[ $? -eq 0 ] || exit 1
+
+	cd ${MESA_OUT_DIR}
+
+	# NOTE: only need vc4 and v3d, so remove other libraries
 	mv ./dist/usr/lib/dri/vc4_dri.so ./dist/usr/lib/dri/vc4_dri.so.tmp
+	mv ./dist/usr/lib/dri/v3d_dri.so ./dist/usr/lib/dri/v3d_dri.so.tmp
 	rm -f ./dist/usr/lib/dri/*_dri.so
 	mv ./dist/usr/lib/dri/vc4_dri.so.tmp ./dist/usr/lib/dri/vc4_dri.so
+	mv ./dist/usr/lib/dri/v3d_dri.so.tmp ./dist/usr/lib/dri/v3d_dri.so
 
 	echo "Done."
 }
@@ -170,13 +166,25 @@ mesa_deb_pkg()
 
 	mkdir ${MESA_DEB_DIR}/DEBIAN
 
+	cat <<-EOF > ${MESA_DEB_DIR}/DEBIAN/rules
+#!/usr/bin/make -f
+
+%:
+	dh $@
+
+override_dh_builddeb:
+	dh_builddeb -- -Zxz
+EOF
+
 	cat <<-EOF > ${MESA_DEB_DIR}/DEBIAN/control
-Package: $MESA_DEB_PKG
-Version: $MESA_DEB_PKG_VER
-Maintainer: $MAINTAINER_NAME <$MAINTAINER_EMAIL>
+Package: ${MESA_DEB_PKG}
+Version: ${MESA_DEB_PKG_VER}
+Maintainer: ${MAINTAINER_NAME} <${MAINTAINER_EMAIL}>
 Architecture: all
 Priority: optional
-Description: This package provides Mesa 3D libraries for sunXi platforms
+Depends: libexpat1 (>= 2.2.0), zlib1g (>= 1.2.0)
+Build-Depends: libexpat1-dev, zlib1g-dev
+Description: This package provides Mesa 3D libraries for Broadcom BCM283x-family SoCs
 EOF
 
 	cat <<-EOF > ${MESA_DEB_DIR}/DEBIAN/postinst
@@ -199,7 +207,7 @@ EOF
 	mkdir -p ${MESA_DEB_DIR}${MESA_PREFIX}
 	rsync -az ${MESA_OUT_DIR}/dist${MESA_PREFIX}/  ${MESA_DEB_DIR}${MESA_PREFIX}
 
-	dpkg-deb -z0 -b $MESA_DEB_DIR  ${BASEDIR}/debs/${MESA_DEB_PKG}.deb
+	dpkg-deb -Zxz -z0 -b ${MESA_DEB_DIR}  ${BASEDIR}/debs/${MESA_DEB_PKG}.deb
 	[ $? -eq 0 ] || exit $?;
 
 	rm -rf ${MESA_DEB_DIR}
@@ -216,7 +224,7 @@ mesa_deploy()
 
 	mkdir -p ${SYSROOT_DIR}${MESA_PREFIX}
 	rsync -az ${MESA_DEB_DIR}${MESA_PREFIX}/  ${SYSROOT_DIR}${MESA_PREFIX}
-	${LIBDIR}/make-relativelinks.sh  $SYSROOT_DIR
+	${LIBDIR}/make-relativelinks.sh  ${SYSROOT_DIR}
 	rm -rf ${MESA_DEB_DIR}
 
 	cp ${BASEDIR}/debs/${MESA_DEB_PKG}.deb  ${R}/tmp/
@@ -228,13 +236,13 @@ mesa_deploy()
 
 # -----------------------------------------------------------------------------
 
-if [ "${INSTALL_MESA}" != "no" ] ; then
+if [ "${ENABLE_MESA}" = yes ] ; then
 
 	echo "Building Mesa..."
 
 	mesa_prepare
 
-	if [[ $CLEAN =~ (^|,)(mesa|mesa-vc4)(,|$) ]] ; then
+	if [[ ${CLEAN} =~ (^|,)(mesa|mesa-vc4)(,|$) ]] ; then
 		rm -f ${BASEDIR}/debs/${MESA_DEB_PKG}.deb
 	fi
 
@@ -247,5 +255,5 @@ if [ "${INSTALL_MESA}" != "no" ] ; then
 
 	echo "Mesa build complete."
 else
-	echo "Skip building MESA library."
+	echo "Skip building MESA (vc4 backend) library."
 fi
